Added swagger docs and Makefile

+++ fullstack/tests/controllertests/post_controller_test.go
@@ -114,7 +114,7 @@ func TestCreatePost(t *testing.T) {
 		if v.statusCode == 201 {
 			assert.Equal(t, responseMap["title"], v.title)
 			assert.Equal(t, responseMap["content"], v.content)
-			assert.Equal(t, responseMap["author_id"], v.author_id)
+			assert.Equal(t, responseMap["author_id"], float64(v.author_id))
 		}
 		if v.statusCode == 401 || v.statusCode == 422 || v.statusCode == 500 && v.errorMessage != "" {
 			assert.Equal(t, responseMap["error"], v.errorMessage)
